requirements.txt
django==1.7
django-mockups
django-suit
djangorestframework